% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlabasterParam.R
\name{AlabasterParam}
\alias{AlabasterParam}
\title{Store MS data objects using the alabaster framework}
\description{
The \href{https://github.com/ArtifactDB/alabaster.base}{\emph{alabaster} framework}
provides the methodology to save R objects to on-disk representations/
storage modes which are programming language independent (in contrast to
e.g. R's RDS files). By using standard file formats such as JSON and HDF5,
alabaster ensures that the data can also be read and imported by other
programming languages such as Python or Javascript. This improves
interoperability between application ecosystems.
}
