% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\docType{data}
\name{all_params}
\alias{all_params}
\title{all parameters for inference}
\format{
An object of class \code{list} of length 17.
}
\usage{
all_params
}
\description{
all inference parameters including immunity
}
\keyword{datasets}
