% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\docType{data}
\name{basic_params}
\alias{basic_params}
\title{basic model parameters}
\format{
An object of class \code{list} of length 1.
}
\usage{
basic_params
}
\description{
parameters for simple global runs
}
\keyword{datasets}
