% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\docType{data}
\name{extended_params}
\alias{extended_params}
\title{extended parameters for inference}
\format{
An object of class \code{list} of length 7.
}
\usage{
extended_params
}
\description{
parameters for inference on:
\itemize{
\item heterogeneity
\item infectivity towards mosquitoes
\item duration of subpatent infection
}
}
\keyword{datasets}
