% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaboLightsParam.R, R/MsExperiment.R
\name{MetaboLightsParam}
\alias{MetaboLightsParam}
\alias{readMsObject,MsExperiment,MetaboLightsParam-method}
\title{Load content from a MetaboLights study}
\usage{
MetaboLightsParam(studyId = character(1))

\S4method{readMsObject}{MsExperiment,MetaboLightsParam}(object, param, ...)
}
\arguments{
\item{studyId}{\code{character(1)} The MetaboLights study studyId, which should start with
"MTBL". This identifier uniquely specifies the study within the
MetaboLights database.}

\item{object}{for \code{saveMsObject()}: the MS data object to save, for
\code{readMsObject()}: the MS data object that should be returned}

\item{param}{an object defining and (eventually configuring) the file format
and file name or directory to/from which the data object should be
exported/imported.}

\item{...}{additional optional arguments. See documentation of respective
method for more information.}
}
\value{
(for now ?) A MsExperiemtn object with only the sampleData slots
filled (will be updated when MetaboLightsBackend available ?).
}
\description{
The \code{MetaboLightsParam} class and the associated \code{readMsObject()} method
allow users to load an \code{MsExperiment} object from a study in the
MetaboLights database (https://www.ebi.ac.uk/metabolights/index) by
providing its unique study \code{studyId}. This function is particularly useful for
importing metabolomics data into an \code{MsExperiment} object for further
analysis within the R environment.

If the study contains multiple assays, the user will be prompted to select
which assay to load. The resulting \code{MsExperiment} object will include a
\code{sampleData} slot populated with data extracted from the selected assay.
Columns in the \code{sampleData} that contain only \code{NA} values are automatically
removed, and an additional column is added to track the injection index.
}
\examples{
library(MsExperiment)
# Load a study with the studyId "MTBLS10035"
param <- MetaboLightsParam(studyId = "MTBLS10035")
ms_experiment <- readMsObject(MsExperiment(), param)

}
\seealso{
\itemize{
\item \code{MsExperiment} object, defined in the
(\href{https://bioconductor.org/packages/MsExperiment}{MsExperiment}) package.
\item \href{https://www.ebi.ac.uk/metabolights/index}{MetaboLights} for accessing
the MetaboLights database.
}
}
\author{
Philippine Louail
}
\concept{MS object export and import formats}
