% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzTabParam.R, R/XcmsExperiment.R
\name{mzTabParam}
\alias{mzTabParam}
\alias{saveMsObject,XcmsExperiment,mzTabParam-method}
\title{Store xcms preprocessing results to a file in mzTab-M format}
\usage{
mzTabParam(
  studyId = character(),
  polarity = c("positive", "negative"),
  sampleDataColumn = character(),
  path = tempdir(),
  optionalFeatureColumns = character(),
  ...
)

\S4method{saveMsObject}{XcmsExperiment,mzTabParam}(object, param)
}
\arguments{
\item{studyId}{\code{character(1)} Will be both the \code{filename} of the object
saved in mzTab-M format and the \code{mzTab-ID} in the file.}

\item{polarity}{\code{character(1)} Describes the polarity of the experiment. Two
inputs are possible, "positive" (default) or "negative".}

\item{sampleDataColumn}{\code{character} with the column name(s) of the
\code{sampleData()} of the \code{XcmsExperiment} object that should be exported
to the mzTab-M file. Defaults to \code{sampleDataColumn = character()}. At
least one column in \code{sampleData()} has to be specified.}

\item{path}{\code{character(1)} Define where the file is going to be stored. The
default is \code{paht = tempdir()}.}

\item{optionalFeatureColumns}{\code{character} with optional column names from
\code{featureDefinitions()} that should be exported. Defaults to
\code{optionalFeatureColumns = character()}. Feature columns \code{"mzmed"},
\code{"rtmed"}, \code{"rtmin"} and \code{"rtmax"} are always exported.}

\item{...}{additional optional arguments. See documentation of respective
method for more information.}

\item{object}{for \code{saveMsObject()}: the MS data object to save, for
\code{readMsObject()}: the MS data object that should be returned}

\item{param}{an object defining and (eventually configuring) the file format
and file name or directory to/from which the data object should be
exported/imported.}
}
\description{
The \code{saveMsObject()} and \code{readMsObject()} methods with the \code{mzTabParam}
option enable users to save/load \code{XcmsExperiment} objects in Mz-Tab-m
file format. Mainly the metadata (MTD) and Small molecule feature (SMF)
tables will represent the \code{XcmsExperiment}. More specifically, \code{sampleData()}
of the object will be stored in the metadata section (MTD) along with the
user-inputed \code{studyId}  and \code{polarity}. The \code{featureDefinitions()} will be
stored in the small molecule feature (SMF) section but by default only the
\code{mzmed}, \code{rtmed}, \code{rtmin} and \code{rtmax} are exported. More info avaialble in
\code{featureDefinitions()} can be exported by specifying the
\code{optionalFeatureColumns} parameter. The \code{featureValues()} will also be
stored in the small molecule feature (SMF) section.

The small molecule summary section (SML) will be filled with \code{null} values
as no annotation and identification of compound is performed in \code{xcms}.

Writing data to a folder that contains already exported data will result in
an error.
}
\section{Slots}{

\describe{
\item{\code{dots}}{Correspond to any optional parameters to be passed
to the \code{featureValues()} function. (e.g. parameters \code{method} or \code{value}).}
}}

\note{
This function was build so that the output fit the recommendation of
mzTab-M file format version 2.0. These can be found \href{http://hupo-psi.github.io/mzTab/2_0-metabolomics-release/mzTab_format_specification_2_0-M_release.html}{here}.
}
\examples{
## Load a test data set with detected peaks, of class `XcmsExperiment`
library(xcms)
test_xcms <- loadXcmsData()

## Define param
param <- mzTabParam(studyId = "test",
                    polarity = "positive",
                    sampleDataColumn = "sample_type")

## Save as a mzTab-M file
saveMsObject(test_xcms, param)

}
\references{
Hoffmann N, Rein J, Sachsenberg T, Hartler J, Haug K, Mayer G, Alka O,
Dayalan S, Pearce JTM, Rocca-Serra P, Qi D, Eisenacher M, Perez-Riverol Y,
Vizcaino JA, Salek RM, Neumann S, Jones AR. mzTab-M: A Data Standard for
Sharing Quantitative Results in Mass Spectrometry Metabolomics. Anal Chem.
2019 Mar 5;91(5):3302-3310. doi: 10.1021/acs.analchem.8b04310. Epub 2019 Feb
13. PMID: 30688441; PMCID: PMC6660005.
}
\seealso{
Other MS object export and import formats.: 
\code{\link{AlabasterParam}},
\code{\link{PlainTextParam}}
}
\author{
Philippine Louail, Johannes Rainer
}
\concept{MS object export and import formats.}
