% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlainTextParam.R, R/MsBackendMzR.R,
%   R/MsExperiment.R, R/Spectra.R, R/XcmsExperiment.R
\name{PlainTextParam}
\alias{PlainTextParam}
\alias{saveMsObject,MsBackendMzR,PlainTextParam-method}
\alias{readMsObject,MsBackendMzR,PlainTextParam-method}
\alias{saveMsObject,MsExperiment,PlainTextParam-method}
\alias{readMsObject,MsExperiment,PlainTextParam-method}
\alias{saveMsObject,Spectra,PlainTextParam-method}
\alias{readMsObject,Spectra,PlainTextParam-method}
\alias{saveMsObject,XcmsExperiment,PlainTextParam-method}
\alias{readMsObject,XcmsExperiment,PlainTextParam-method}
\title{Store contents of MS objects as plain text files}
\usage{
PlainTextParam(path = tempdir())

\S4method{saveMsObject}{MsBackendMzR,PlainTextParam}(object, param)

\S4method{readMsObject}{MsBackendMzR,PlainTextParam}(object, param, spectraPath = character())

\S4method{saveMsObject}{MsExperiment,PlainTextParam}(object, param)

\S4method{readMsObject}{MsExperiment,PlainTextParam}(object, param, spectraPath = character())

\S4method{saveMsObject}{Spectra,PlainTextParam}(object, param)

\S4method{readMsObject}{Spectra,PlainTextParam}(object, param, spectraPath = character())

\S4method{saveMsObject}{XcmsExperiment,PlainTextParam}(object, param)

\S4method{readMsObject}{XcmsExperiment,PlainTextParam}(object, param, spectraPath = character())
}
\arguments{
\item{path}{For \code{PlainTextParam()}: \code{character(1)}, defining where the files
are going to be stored/ should be loaded from. The default is
\code{path = tempdir()}.}

\item{object}{for \code{saveMsObject()}: the MS data object to save, for
\code{readMsObject()}: the MS data object that should be returned}

\item{param}{an object defining and (eventually configuring) the file format
and file name or directory to/from which the data object should be
exported/imported.}

\item{spectraPath}{For \code{readMsObject()}: \code{character(1)} optionally allowing to
define the (absolute) path where the spectra files (\emph{data storage files})
can be found. This parameter is passed to the \code{loadResults()} method of
the MsBackend().}
}
\value{
For \code{PlainTextParam()}: a \code{PlainTextParam} class. \code{saveMsObject()}
does not return anything but saves the object to collections of different
plain text files to a folder. The \code{readMsObject()} method returns the
restored data as an instance of the class specified with parameter \code{object}.
}
\description{
The \code{saveMsObject()} and \code{readMsObject()} methods with the \code{PlainTextParam}
option enable users to save/load different type of MS object as a
collections of plain text files in/from a specified folder. This folder,
defined with the \code{path} parameter, will be created by the \code{storeResults()}
function. Any previous exports eventually present in that folder will be
overwritten.

The MS object currently supported for import and export with this parameter
are :
\itemize{
\item \code{MsBackendMzR} object
\item \code{Spectra} object
\item \code{MsExperiment} object
\item \code{XcmsExperiment} object
}
}
\author{
Philippine Louail
}
\concept{MS object export and import formats.}
