% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_simulations_from_data}
\alias{run_simulations_from_data}
\title{Run simulations from data}
\usage{
run_simulations_from_data(
  node,
  warmup,
  paramset,
  seed,
  datadir,
  batch_size,
  n_batches,
  outdir
)
}
\arguments{
\item{node}{an arbitrary number, useful for tracking multiple executions}

\item{warmup}{number of years to warm up for}

\item{paramset}{a list of parameters from sample.R}

\item{seed}{random seed}

\item{datadir}{directory for global data to sample from}

\item{batch_size}{number of runs per batch}

\item{n_batches}{number of batches}

\item{outdir}{directory to save outputs}
}
\description{
Run simulations from data
}
