% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_synthetic_simulations}
\alias{run_synthetic_simulations}
\title{Run synthetic simulations}
\usage{
run_synthetic_simulations(
  node = 1,
  n_years = 5,
  warmup = 5,
  reps = 10,
  paramset = basic_params,
  seed = 42,
  sites = NULL,
  interventions = "nets",
  batch_size = 1,
  n_batches = 1,
  outdir = ".",
  outputs = "prev",
  aggregation = "daily",
  synthetic_intervention_method = "lhs",
  human_population = 1e+05,
  seasonality_output = "daily"
)
}
\arguments{
\item{node}{an arbitrary number, useful for tracking multiple executions}

\item{n_years}{number of years to simulate}

\item{warmup}{number of years to warm up for}

\item{paramset}{a list of parameters from sample.R}

\item{seed}{random seed}

\item{interventions}{vector of interventions to include}

\item{batch_size}{number of runs per batch}

\item{n_batches}{number of batches}

\item{outdir}{directory to save outputs}

\item{outputs}{character vector of outputs to include}

\item{aggregation}{type of aggregation for outputs, either 'daily' or 'yearly'}
}
\description{
Run synthetic simulations
}
