% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{saveMsObject}
\alias{saveMsObject}
\alias{saveMsObject,MsBackend,ANY-method}
\alias{loadMsObject,MsBackend,ANY-method}
\alias{loadMsObject}
\title{Save and load MS data objects to and from different file formats}
\usage{
saveMsObject(object, param, ...)

loadMsObject(object, param, ...)
}
\arguments{
\item{object}{for \code{saveMsObject()}: the MS data object to save, for
\code{loadMsObject()}: the MS data object that should be returned}

\item{param}{an object defining and (eventually configuring) the file format
and file name or directory to/from which the data object should be
exported/imported.}

\item{...}{additional optional arguments. See documentation of respective
method for more information.}
}
\value{
\code{saveMsObject()} has no return value, \code{loadMsObject} is expected
to return an instance of the class defined with \code{object}.
}
\description{
The \code{saveMsObject()} and \code{loadMsObject()} methods allow serializing and
restoring/importing mass spectrometry (MS) data objects to and from language
agnostic file formats. The type  and configuration of the file format is
defined by the second argument to the method, \code{param}.
\itemize{
\item \code{saveMsObject(object, param)}: saves the MS data object \code{object} to file(s)
in a format defined by \code{param}.
\item \code{loadMsObject(object, param)}: \code{object} defines the type of MS object that
should be returned by the function and \code{param} the format and file name(s)
from which the data should be restored/imported.
}
}
\author{
Philippine Louail, Johannes Rainer, Laurent Gatto
}
