% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaboLightsParam.R, R/MsExperiment.R
\name{MetaboLightsParam}
\alias{MetaboLightsParam}
\alias{readMsObject,MsExperiment,MetaboLightsParam-method}
\title{Load content from a MetaboLights study}
\usage{
MetaboLightsParam(
  mtblsId = character(),
  assayName = character(),
  filePattern = "mzML$|CDF$|cdf$|mzXML$"
)

\S4method{readMsObject}{MsExperiment,MetaboLightsParam}(
  object,
  param,
  keepOntology = TRUE,
  keepProtocol = TRUE,
  simplify = TRUE,
  ...
)
}
\arguments{
\item{mtblsId}{\code{character(1)} The MetaboLights study ID, which should
start with "MTBL". This identifier uniquely specifies the study within the
MetaboLights database.}

\item{assayName}{\code{character(1)} The name of the assay to load. If the study
contains multiple assays and this parameter is not specified, the user will
be prompted to select which assay to load.}

\item{filePattern}{\code{character(1)} A regular expression pattern to filter the
raw data files associated with the selected assay. The default value is
\code{"mzML$|CDF$|cdf$|mzXML$"}, corresponding to the supported raw data file
types.}

\item{object}{for \code{saveMsObject()}: the MS data object to save, for
\code{readMsObject()}: the MS data object that should be returned}

\item{param}{an object defining and (eventually configuring) the file format
and file name or directory to/from which the data object should be
exported/imported.}

\item{keepOntology}{\code{logical(1)} Whether to keep columns related to ontology
in the \code{sampleData} parameter. Default is \code{TRUE}.}

\item{keepProtocol}{\code{logical(1)} Whether to keep columns related to protocols
information in the \code{sampleData} parameter. Default is \code{TRUE}.}

\item{simplify}{\code{logical(1)} Whether to simplify the \code{sampleData} table by
removing columns filled with NAs or duplicated content. Default is \code{TRUE}.}

\item{...}{additional optional arguments. See documentation of respective
method for more information.}
}
\value{
An \code{MsExperiment} object with the \code{sampleData} parameter populated using
MetaboLights sample and assay information. The spectra data is represented
as a \code{MsBackendMetabolights} object, generated from the raw data files
associated with the selected assay of the specified MetaboLights ID
(\code{mtblsId}).
}
\description{
The \code{MetaboLightsParam} class and the associated \code{readMsObject()} method
allow users to load an \code{MsExperiment} object from a study in the
MetaboLights database (https://www.ebi.ac.uk/metabolights/index) by
providing its unique study \code{mtblsId}. This function is particularly useful
for importing metabolomics data into an \code{MsExperiment} object for further
analysis in the R environment.
It is important to note that at present it is only possible to \emph{read}
(import) data from MetaboLights, but not to \emph{save} data to MetaboLights.

If the study contains multiple assays, the user will be prompted to select
which assay to load. The resulting \code{MsExperiment} object will include a
\code{sampleData} slot populated with data extracted from the selected assay.

Users can define how to filter this \code{sampleData} table by specifying a few
parameters. The \code{keepOntology} parameter is set to \code{TRUE} by default, meaning
that all ontology-related columns are retained. If set to \code{FALSE}, they are
removed. If ontology columns are kept, some column names may be duplicated and
therefore numbered. The order of these columns is important, as it reflects the
assay and sample information available in MetaboLights.

The \code{keepProtocol} parameter is also set to \code{TRUE} by default, meaning that
all columns related to protocols are kept. If set to \code{FALSE}, they are removed.
The \code{simplify} parameter (default \code{simplify = TRUE}) allows to define
whether duplicated columns or columns containing only missing values should
be removed. In the case of duplicated content, only the first occurring
column will be retained.

Further filtering can be performed using the \code{filePattern} parameter of the
\code{MetaboLightsParam} object. The default for this parameter is
\code{"mzML$|CDF$|cdf$|mzXML$"}, which corresponds to the supported raw data file
types.
}
\examples{

library(MsExperiment)
# Load a study with the mtblsId "MTBLS39" and selecting specific file pattern
# as well as removing ontology and protocol information in the metadata.
param <- MetaboLightsParam(mtblsId = "MTBLS39", filePattern = "63A.cdf")
ms_experiment <- readMsObject(MsExperiment(), param , keepOntology = FALSE,
                              keepProtocol = FALSE)

}
\seealso{
\itemize{
\item \code{MsExperiment} object, defined in the
(\href{https://bioconductor.org/packages/MsExperiment}{MsExperiment}) package.
\item \code{MsBackendMetaboLights} object, defined in the
(\href{https://github.com/rformassspectrometry/MsBackendMetaboLights}{MsBackendMetaboLights})
repository.
\item \href{https://www.ebi.ac.uk/metabolights/index}{MetaboLights} for accessing
the MetaboLights database.
}
}
\author{
Philippine Louail
}
\concept{MS object export and import formats}
