% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MzTabParam.R, R/XcmsExperiment.R
\name{MzTabParam}
\alias{MzTabParam}
\alias{saveMsObject,XcmsExperiment,MzTabParam-method}
\title{Store contents of an \code{XcmsExperiment} objects as a .mztabm file}
\usage{
MzTabParam(
  studyId = character(),
  polarity = c("positive", "negative"),
  sampleDataColumn = character(),
  path = tempdir(),
  optionalFeatureColumns = character(),
  ...
)

\S4method{saveMsObject}{XcmsExperiment,MzTabParam}(object, param)
}
\arguments{
\item{studyId}{\code{character(1)} Will be both the \code{filename} of the object
saved in .mztabm format and the \code{mzTab-ID} in the file.}

\item{polarity}{\code{character(1)} Describes the polarity of the experiment. Two
inputs are possible, "positive" (default) or "negative".}

\item{sampleDataColumn}{\code{character} strings corresponding to the column
name(s) of the \code{sampleData()} of the \code{XcmsExperiment} object with the
different \emph{variables} of the experiment, for example it could be
\emph{"phenotype"}, \emph{"sample_type"}, etc...}

\item{path}{\code{character(1)} Define where the file is going to be stored. The
default will be \code{tempdir()}.}

\item{optionalFeatureColumns}{Optional columns from \code{featureDefinitions}
that should be exported too. For example it could be \emph{"peakidx},
\emph{"npeaks"}, etc...}

\item{...}{additional optional arguments. See documentation of respective
method for more information.}

\item{object}{for \code{saveMsObject()}: the MS data object to save, for
\code{readMsObject()}: the MS data object that should be returned}

\item{param}{an object defining and (eventually configuring) the file format
and file name or directory to/from which the data object should be
exported/imported.}
}
\description{
The \code{saveMsObject()} and \code{readMsObject()} methods with the \code{MzTabParam}
option enable users to save/load \code{XcmsExperiment} objects in/from a .mztabm
file format. Mainly the metadata (MTD) and Small molecule feature (SMF)
tables will represent the \code{XcmsExperiment}. The small molecule summary
section (SML) will be filled with \code{null} values as no annotation and
identification of compound is performed in \code{xcms}. Writing data to a folder
that contains already exported data will result in an error.
}
\section{Slots}{

\describe{
\item{\code{dots}}{\code{list} Correspond to any optional parameters to be passed
to the \code{featureValues} function. (e.g. parameters \code{method} or \code{value}).}
}}

\note{
This function was build so that the output fit the recommendation of mztab-m
file format. These can be found here:
(http://hupo-psi.github.io/mzTab/2_0-metabolomics-release/mzTab_format_specification_2_0-M_release.html)
}
\examples{
## Load a test data set with detected peaks, of class `XcmsExperiment`
library(xcms)
test_xcms <- loadXcmsData()

## Define param
param <- MzTabParam(studyId = "test",
                    polarity = "positive",
                    sampleDataColumn = "sample_type")

## Save as a .mzTabm file
saveMsObject(test_xcms, param)

}
\references{
Hoffmann N, Rein J, Sachsenberg T, Hartler J, Haug K, Mayer G, Alka O,
Dayalan S, Pearce JTM, Rocca-Serra P, Qi D, Eisenacher M, Perez-Riverol Y,
Vizcaino JA, Salek RM, Neumann S, Jones AR. mzTab-M: A Data Standard for
Sharing Quantitative Results in Mass Spectrometry Metabolomics. Anal Chem.
2019 Mar 5;91(5):3302-3310. doi: 10.1021/acs.analchem.8b04310. Epub 2019 Feb
13. PMID: 30688441; PMCID: PMC6660005.
}
\seealso{
Other MS object export and import formats.: 
\code{\link{AlabasterParam}},
\code{\link{PlainTextParam}}
}
\author{
Philippine Louail, Johannes Rainer
}
\concept{MS object export and import formats.}
